newTriggerName, setNewTriggerName and addTrigger added to context and imported into NewTriggerInput
