+++ src/app/components/new-trigger-input.tsx
@@ -1,22 +1,10 @@
 "use client";
 
-import { useState } from "react";
 import { useTriggerContext } from "@/context/trigger";
 
 export default function NewTriggerInput() {
-  const [newTriggerName, setNewTriggerName] = useState("");
-  const { setTriggers } = useTriggerContext();
+  const { newTriggerName, setNewTriggerName, addTrigger } = useTriggerContext();
 
-  // TODO: This should be a POST call to the server.
-  const addTrigger = () => {
-    if (newTriggerName.trim() !== "") {
-      setTriggers((prev) => [
-        ...prev,
-        { name: newTriggerName, successCount: 0, failureCount: 0, winRate: 0 },
-      ]);
-      setNewTriggerName("");
-    }
-  };
   return (
     <div className="mb-4 flex gap-2">
       <input
