+++ src/context/trigger.tsx
@@ -7,7 +7,9 @@ import { Trigger } from "@/app/lib/types";
 type TriggerContext = {
   triggers: Trigger[];
   setTriggers: React.Dispatch<React.SetStateAction<Trigger[]>>;
-  
+  newTriggerName: string;
+  setNewTriggerName: React.Dispatch<React.SetStateAction<string>>;
+  addTrigger: () => void;
 };
 
 export const TriggerContext = createContext<TriggerContext | null>(null);
@@ -18,15 +20,35 @@ export default function TriggerContextProvider({
   children: React.ReactNode;
 }) {
   const [triggers, setTriggers] = useState<Trigger[]>([]);
+  const [newTriggerName, setNewTriggerName] = useState("");
   const [isLoaded, setIsLoaded] = useState(false);
 
+  // Make sire this is necessary.
   useEffect(() => {
     setIsLoaded(true);
   }, []);
-  
+
+  // TODO: This should be a POST call to the server.
+  const addTrigger = () => {
+    if (newTriggerName.trim() !== "") {
+      setTriggers((prev) => [
+        ...prev,
+        { name: newTriggerName, successCount: 0, failureCount: 0, winRate: 0 },
+      ]);
+      setNewTriggerName("");
+    }
+  };
 
   return (
-    <TriggerContext.Provider value={{ triggers, setTriggers }}>
+    <TriggerContext.Provider
+      value={{
+        triggers,
+        setTriggers,
+        newTriggerName,
+        setNewTriggerName,
+        addTrigger,
+      }}
+    >
       {isLoaded ? children : null}
     </TriggerContext.Provider>
   );
@@ -35,7 +57,9 @@ export default function TriggerContextProvider({
 export const useTriggerContext = () => {
   const context = useContext(TriggerContext);
   if (!context) {
-    throw new Error("useTriggerContext must be used within a TriggerContextProvider");
+    throw new Error(
+      "useTriggerContext must be used within a TriggerContextProvider"
+    );
   }
   return context;
 };
